Add latex env support
TODO: This is a super janky way of supporting latex. We first
parse \begin{...} \end{...} into a raw literal

Then later, inside slate, we rerun the parser on the raw literal
to get any nested markdown.

This is def. slower than doing it "correctly" but implementation
complexity is much lower